Set up test database and remove copied over setup that we might not need yet.
- Since we copied this code over from the previous Mosaic Platform API,
we've updated the setup for the test database used for TDD and End-to-End (Integration) testing to
match our current usecase.

- This also removes the Amazon XRay Tracing set up as this something that hindered creating the migrations.

- There might be a need to discuss whether or not we need to have the migrations set up as it came up that we might
not need it right now while we currently use DMS to grab the tables we need from the recovered Mosaic Database.

- We don't need the XRay tracing yet but would be good to understand what it does and how
to set it up for our project specifically rather than just keeping the setup used for the Mosaic Platform API.