Add list case note gateway method
This creates a gateway method that when called will query the database and
return a list of objects with some of their fields populated.

Only some fields are populated per object as the UI where this LIST endpoint will be called
displays summary fields of each record a person has.