Update list case note gateway to return more information about a case note
The list case note gateway was initially created to return a few properties describing a case note.
This was done so that we could get the list case note endpoint tested and working end to end quicker.

This change modifies the list case note endpoint to return everything about a case note apart from the
case note content -> This is returned when viewing a specific case note.